Scope the persistence claims to window placement; qualify same-id semantics
Round-11 review: 1 comment, 4 suppressed.

Overclaim, and a real one. The CHANGELOG said this store makes "window/dock
layouts" survive a rename, but docking does not use IWindowPersistenceStore
at all -- DockManager.PersistenceId stores its JSON under
WindowPersistedScope["docking:<id>"], an IPersistedStateScope. Verified in
source before changing. Nothing in this PR affects dock persistence, so the
changelog, and spec 063's motivation section, now say window placement and
state the distinction explicitly.

Concurrency wording was also too strong in the guide and the shipped skill.
Serializing writes removes the LOST-UPDATE race between different
persistence ids; it does not make two windows that share one id
independently persistent -- that remains last-writer-wins on a single key,
by design. Both now say so.

Null-product case added: the constructor guards publisher and product
symmetrically but only the publisher null was pinned, so a regression that
dropped the product guard stayed green.

Disputed finding, recorded for the record: the review flagged
`new byte[JsonFileStore.MaxFileSizeBytes]` as a compile error because the
constant is a long. It is not -- C# accepts a long array bound, and the
build has been green across the last several rounds with that exact line,
including the run that added it. I added the (int) cast anyway for
consistency with the neighbouring oversize-read case, not to fix a defect.

Reactor.Tests 14056/0; Release build matches the pre-change baseline; guide
regenerated.

+++ tests/Reactor.Tests/Diagnostics/PersistenceEtwBridgeTests.cs
@@ -213,8 +213,10 @@ public class PersistenceEtwBridgeTests : IDisposable
         try
         {
             // Exceeds JsonFileStore.MaxFileSizeBytes once base64-encoded, so the write
-            // takes the oversize-rejection arm.
-            store.Write("main", new byte[JsonFileStore.MaxFileSizeBytes]);
+            // takes the oversize-rejection arm. Cast for consistency with the
+            // oversize-read case below; the constant is a long, which C# accepts as an
+            // array bound either way.
+            store.Write("main", new byte[(int)JsonFileStore.MaxFileSizeBytes]);
 
             AssertEvent(
                 _listener.Events,
